requirements.txt
djangorestframework
djangorestframework-simplejwt
drf-yasg
